add conf to install on Amazon rds

+++ src/main/java/in/bushansirgur/expensetrackerapi/repository/ExpenseRepository.java
@@ -14,17 +14,17 @@ import java.util.Optional;
 public interface ExpenseRepository extends JpaRepository<Expense, Integer> {
 
     // SELECT * FROM tbl_expenses WHERE category='Other'
-    Page<Expense> findByUserIdAndCategory(Long userId,String category, Pageable page);
+    Page<Expense> findByUserIdAndCategory(Integer userId,String category, Pageable page);
 
     // SELECT * FROM tbl_expenses WHERE expense_name Like '%a%'
-    Page<Expense> findByUserIdAndNameContaining(Long userId,String keyword, Pageable page);
+    Page<Expense> findByUserIdAndNameContaining(Integer userId,String keyword, Pageable page);
 
     // SELECT * FROM tbl_expenses WHERE date BETWEEN date('2021-10-13') AND date('2021-10-15')
-    Page<Expense> findByUserIdAndDateBetween(Long userId,LocalDate startDate, LocalDate endDate, Pageable page);
+    Page<Expense> findByUserIdAndDateBetween(Integer userId,LocalDate startDate, LocalDate endDate, Pageable page);
 
-    Page<Expense> findByUserId(Long userId, Pageable page);
+    Page<Expense> findByUserId(Integer userId, Pageable page);
 
-    Optional<Expense> findByUserIdAndId(Long userId, Integer expenseId);
+    Optional<Expense> findByUserIdAndId(Integer userId, Integer expenseId);
 
 
 }
